+++ addon/components/object-list-view.js
@@ -650,8 +650,11 @@ export default FlexberryBaseComponent.extend(
       @param {jQuery.Event} e jQuery.Event by click on row
     */
     rowClick(recordWithKey, e) {
+      let editOnSeparateRoute = this.get('editOnSeparateRoute');
       if (this.get('readonly')) {
-        return;
+        if (!editOnSeparateRoute) {
+          return;
+        }
       }
 
       if (this.rowClickable) {
@@ -667,6 +670,7 @@ export default FlexberryBaseComponent.extend(
           editOnSeparateRoute: editOnSeparateRoute,
           modelName: this.get('modelProjection').modelName,
           detailArray: this.get('content'),
+          readonly: this.get('readonly')
         });
       }
     },